requirements.txt
# Основные зависимости
python-telegram-bot==20.8
python-dotenv==1.0.1
nest-asyncio==1.6.0

# Логирование
logging==0.4.9.6
